make project work

+++ Poseiden-skeleton/src/main/java/com/nnk/springboot/repositories/BidListRepository.java
@@ -4,6 +4,6 @@ import com.nnk.springboot.domain.BidList;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 
-public interface BidListRepository extends JpaRepository<BidList, Integer> {
+public interface BidListRepository extends JpaRepository<BidList, Long> {
 
 }
